+++ speculos/main.py
@@ -50,8 +50,9 @@ class ElfInfo:
     fonts_size: int = 0
     app_nvram_addr: int = 0
     app_nvram_size: int = 0
-    cx_ram_addr: int = 0
-    cx_ram_size: int = 0
+    shared_ram_addr: int = 0
+    shared_ram_size: int = 0
+    pic_init_addr: int = 0
 
 
 DEFAULT_SEED = ('glory promote mansion idle axis finger extra february uncover one trip resource lawn turtle enact '
@@ -140,17 +141,45 @@ def get_elf_infos(app_path, use_bagl, args):
     return ei
 
 
-def get_cx_infos(app_path):
+def get_sharedlib_infos(app_path, apiLevel):
     ei = ElfInfo()
     with open(app_path, 'rb') as fp:
         elf = ELFFile(fp)
         text_section = elf.get_section_by_name('.text')
-        cxram_section = elf.get_section_by_name('.cxram')
+        # The name of the section is shram, starting at API Level 23
+        if apiLevel is None or int(apiLevel) < 23:
+            sharedram_section = elf.get_section_by_name('.cxram')
+        else:
+            sharedram_section = elf.get_section_by_name('.shram')
         ei.text_offset = text_section['sh_offset']
         ei.text_size = text_section['sh_size']
         ei.text_addr = text_section['sh_addr']
-        ei.cx_ram_addr = cxram_section["sh_addr"]
-        ei.cx_ram_size = cxram_section["sh_size"]
+        ei.shared_ram_addr = sharedram_section["sh_addr"]
+        ei.shared_ram_size = sharedram_section["sh_size"]
+        # Look for the symbols SVC_Call and SVC_cx_call
+        # if they are found, save their addresses to patch them to replace the SYSCALL later
+        symtab_section = elf.get_section_by_name('.symtab')
+        if symtab_section is not None:
+            svc_call_symbol = symtab_section.get_symbol_by_name("SVC_Call")
+            if svc_call_symbol is not None:
+                ei.svc_call_addr = svc_call_symbol[0]['st_value'] & (~1)
+            svc_cx_call_symbol = symtab_section.get_symbol_by_name("SVC_cx_call")
+            if svc_cx_call_symbol is not None:
+                ei.svc_cx_call_addr = svc_cx_call_symbol[0]['st_value'] & (~1)
+        # At API Level 23, fonts are stored in shared elf, in C_nbgl_fonts variable
+        if apiLevel is not None and int(apiLevel) >= 23:
+            symtab = elf.get_section_by_name('.symtab')
+            nbgl_fonts_symbol = symtab.get_symbol_by_name('C_nbgl_fonts')
+            if nbgl_fonts_symbol is not None:
+                ei.fonts_addr = nbgl_fonts_symbol[0]['st_value']
+                ei.fonts_size = nbgl_fonts_symbol[0]['st_size']
+                logger.info(f"Found C_nbgl_fonts at 0x{ei.fonts_addr:X} ({ei.fonts_size} bytes)\n")
+            else:
+                logger.info("Disabling OCR.")
+            # At API Level >= 23, a function called pic_init() needs to be retrieved
+            pic_init_symbol = symtab.get_symbol_by_name("pic_init")
+            if pic_init_symbol is not None:
+                ei.pic_init_addr = pic_init_symbol[0]['st_value']
 
     return ei
 
@@ -176,19 +205,29 @@ def run_qemu(s1: socket.socket, s2: socket.socket, args: argparse.Namespace) -> 
     if args.pki_prod:
         argv += ['-p']
 
-    # load cxlib only if available for the specified api level or sdk
+    fonts_addr = 0
+    fonts_size = 0
+
+    # load shared lib only if available for the specified api level or sdk
     if args.apiLevel:
-        cxlib_filepath = f"cxlib/{args.model}-api-level-cx-{args.apiLevel}.elf"
+        if int(args.apiLevel) < 23:
+            sharedlib_filepath = f"cxlib/{args.model}-api-level-cx-{args.apiLevel}.elf"
+        else:
+            sharedlib_filepath = f"sharedlib/{args.model}-api-level-shared-{args.apiLevel}.elf"
     else:
-        cxlib_filepath = f"cxlib/{args.model}-cx-{args.sdk}.elf"
-    cxlib = str(resources.files(__package__) / cxlib_filepath)
-    if os.path.exists(cxlib):
-        ei = get_cx_infos(cxlib)
-        cxlib_args = f'{cxlib}:{ei.text_offset:#x}:{ei.text_size:#x}:{ei.text_addr:#x}'
-        cxlib_args += f':{ei.cx_ram_size:#x}:{ei.cx_ram_addr:#x}'
-        argv += ['-c', cxlib_args]
+        sharedlib_filepath = f"cxlib/{args.model}-cx-{args.sdk}.elf"
+    sharedlib = str(resources.files(__package__) / sharedlib_filepath)
+    if os.path.exists(sharedlib):
+        ei = get_sharedlib_infos(sharedlib, args.apiLevel)
+        sharedlib_args = f'{sharedlib}:{ei.text_offset:#x}:{ei.text_size:#x}:{ei.text_addr:#x}'
+        sharedlib_args += f':{ei.shared_ram_size:#x}:{ei.shared_ram_addr:#x}'
+        sharedlib_args += f':{ei.pic_init_addr:#x}'
+        sharedlib_args += f':{ei.svc_call_addr:#x}:{ei.svc_cx_call_addr:#x}'
+        argv += ['-c', sharedlib_args]
+        fonts_addr = ei.fonts_addr
+        fonts_size = ei.fonts_size
     else:
-        logger.warning(f"Cx lib {cxlib_filepath} not found")
+        logger.warn(f"Shared lib {sharedlib_filepath} not found")
 
     only_bagl = True
     # 'bagl' is the default value of the binary.sections.sdk_graphics. We need to
@@ -206,6 +245,11 @@ def run_qemu(s1: socket.socket, s2: socket.socket, args: argparse.Namespace) -> 
             only_bagl = False
         ei = get_elf_infos(lib_path, use_bagl, args)
 
+        # if fonts_addr and fonts_size have not been found in shared.elf, use the
+        # ones for app.elf (but it was only for BAGL apps)
+        if fonts_addr == 0:
+            fonts_addr = ei.fonts_addr
+            fonts_size = ei.fonts_size
         # Since binaries loaded as libs could also declare extra RAM page(s), collect them all
         if (ei.ram_addr, ei.ram_size) != (0, 0):
             arg = f'{ei.ram_addr:#x}:{ei.ram_size:#x}'
@@ -216,15 +260,15 @@ def run_qemu(s1: socket.socket, s2: socket.socket, args: argparse.Namespace) -> 
         lib_arg = f'{name}:{lib_path}:{ei.text_offset:#x}:{ei.text_size:#x}'
         lib_arg += f':{ei.stack_addr:#x}:{ei.stack_size:#x}:{ei.svc_call_addr:#x}'
         lib_arg += f':{ei.svc_cx_call_addr:#x}:{ei.text_addr:#x}'
-        lib_arg += f':{ei.fonts_addr:#x}:{ei.fonts_size:#x}'
+        lib_arg += f':{fonts_addr:#x}:{fonts_size:#x}'
         lib_arg += f':{ei.app_nvram_addr:#x}:{ei.app_nvram_size:#x}'
         lib_arg += f':{1 if args.load_nvram else 0}'
         lib_arg += f':{1 if args.save_nvram else 0}'
         lib_arg += f':{1 if not use_bagl else 0}'
         argv.append(lib_arg)
 
-    # for NBGL apps, fonts binary file is mandatory
-    if not only_bagl:
+    # for NBGL apps, fonts binary file is mandatory before API Level 23
+    if not only_bagl and int(args.apiLevel) < 23:
         fonts_filepath = f"fonts/{args.model}-fonts-{args.apiLevel}.bin"
         fonts = str(resources.files(__package__) / fonts_filepath)
         if os.path.exists(fonts):