+++ src/bolos/os_types.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <stdbool.h>
+#include <stdint.h>
+
+typedef uint8_t bolos_bool_t;
+
+typedef uint32_t bolos_err_t;
+
+#define BOLOS_TRUE  0xaa
+#define BOLOS_FALSE 0x55