+++ tests/c/syscalls/test_endorsement.c
@@ -163,9 +163,91 @@ void test_endorsement_new(void **state __attribute__((unused)))
                    1);
 };
 
+void test_endorsement_after_api_level_22(void **state __attribute__((unused)))
+{
+  init_environment();
+
+  uint8_t raw_endorsement_pubkey[65] = { 0 };
+  uint8_t endorsement_sig[80] = { 0 };
+  uint8_t endorsement_key1_sig[80] = { 0 };
+  size_t endorsement_key1_sig_len;
+  uint8_t pubkey_len = 0, endorsement_sig_len = 0;
+  cx_ecfp_public_key_t endorsement_pubkey, blue_owner_endorsement_pubkey;
+  cx_sha256_t sha256;
+  uint8_t certif_prefix = 0xFE;
+  uint8_t hash[32];
+  uint8_t code_hash[32];
+  uint8_t nonce[4] = { 0x12, 0x34, 0x56, 0x78 };
+
+  // test key
+  uint8_t BLUE_OWNER_PUBLIC_KEY[] = {
+    0x04, 0x41, 0x14, 0xd1, 0x6a, 0xec, 0xed, 0xa8, 0xd7, 0x81, 0x31,
+    0x1b, 0xd3, 0x4f, 0xd5, 0x7a, 0x44, 0xac, 0xfc, 0xbc, 0xf3, 0x57,
+    0xf0, 0x67, 0x88, 0x46, 0x48, 0x07, 0x32, 0x52, 0x69, 0x9c, 0xbd,
+    0x46, 0xf6, 0x41, 0xc1, 0x75, 0x54, 0xed, 0x65, 0x3e, 0x6a, 0x42,
+    0x7f, 0x4a, 0xea, 0xfa, 0xf7, 0x4e, 0x32, 0x18, 0xe5, 0xdd, 0x0c,
+    0x04, 0x20, 0xb6, 0xc8, 0xcf, 0x60, 0x5a, 0x04, 0x96, 0x4e
+  };
+
+  // -- Check get public key
+  assert_int_equal(0, sys_ENDORSEMENT_get_public_key(ENDORSEMENT_SLOT_1,
+                                                     raw_endorsement_pubkey,
+                                                     &pubkey_len));
+  assert_int_equal(pubkey_len, 65);
+  assert_int_equal(raw_endorsement_pubkey[0], 0x04);
+  assert_int_equal(cx_ecfp_init_public_key(CX_CURVE_256K1,
+                                           raw_endorsement_pubkey, pubkey_len,
+                                           &endorsement_pubkey),
+                   pubkey_len);
+  assert_int_equal(cx_ecfp_init_public_key(CX_CURVE_256K1,
+                                           BLUE_OWNER_PUBLIC_KEY,
+                                           sizeof(BLUE_OWNER_PUBLIC_KEY),
+                                           &blue_owner_endorsement_pubkey),
+                   sizeof(BLUE_OWNER_PUBLIC_KEY));
+
+  // -- Check get public key certificate
+  assert_int_equal(cx_sha256_init(&sha256), CX_SHA256);
+  assert_int_equal(cx_hash(&sha256.header, 0, &certif_prefix, 1, NULL, 0), 0);
+  assert_int_equal(cx_hash(&sha256.header, CX_LAST, raw_endorsement_pubkey,
+                           sizeof(raw_endorsement_pubkey), hash,
+                           CX_SHA256_SIZE),
+                   CX_SHA256_SIZE);
+
+  assert_int_equal(
+      0, sys_ENDORSEMENT_get_public_key_certificate(
+             ENDORSEMENT_SLOT_1, endorsement_sig, &endorsement_sig_len));
+  assert_int_equal(cx_ecdsa_verify(&blue_owner_endorsement_pubkey, CX_LAST,
+                                   CX_SHA256, hash, sizeof(hash),
+                                   endorsement_sig, endorsement_sig_len),
+                   1);
+
+  // -- Check get code hash
+  assert_int_equal(sys_ENDORSEMENT_get_code_hash(code_hash), 0);
+
+  // -- Check key1 signature
+  assert_int_equal(cx_sha256_init(&sha256), CX_SHA256);
+  assert_int_equal(cx_hash(&sha256.header, 0, nonce, sizeof(nonce), NULL, 0),
+                   0);
+  assert_int_equal(cx_hash(&sha256.header, CX_LAST, code_hash,
+                           sizeof(code_hash), hash, CX_SHA256_SIZE),
+                   CX_SHA256_SIZE);
+
+  sys_ENDORSEMENT_key1_sign_data(nonce, sizeof(nonce), endorsement_key1_sig,
+                                 &endorsement_key1_sig_len);
+  assert_int_equal(cx_ecdsa_verify(&endorsement_pubkey, CX_LAST, CX_SHA256,
+                                   hash, sizeof(hash), endorsement_key1_sig,
+                                   endorsement_key1_sig_len),
+                   1);
+};
+
 int main(void)
 {
-  const struct CMUnitTest tests[] = { cmocka_unit_test(test_endorsement),
-                                      cmocka_unit_test(test_endorsement_new) };
+  // clang-format off
+  const struct CMUnitTest tests[] = {
+    cmocka_unit_test(test_endorsement),
+    cmocka_unit_test(test_endorsement_new),
+    cmocka_unit_test(test_endorsement_after_api_level_22)
+  };
+  // clang-format on
   return cmocka_run_group_tests(tests, NULL, NULL);
 }