+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.0] 2025-04-18
+
+### Added
+
+- Add support of API level 23 (Cx -> Shared)
+- Implement new endorsement syscalls"
+- Implement NBGL adaptations to avoid drawing outside of area/screen
+
 ## [0.19.0] 2025-03-31
 
 ### Changed